feat: 🎸 update Text storybook
