+++ libs/ui/src/lib/forms/Button/Button.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Story, Meta } from '@storybook/react/types-6-0';
 
-import { Button } from './Button';
+import { Button, ButtonProps } from './Button';
 import { guiaBolso } from '../../../designSystem/tokens';
 
 export default {
@@ -9,7 +9,7 @@ export default {
   component: Button,
 } as Meta;
 
-const Template: Story<typeof Button> = args => <Button {...args} />;
+const Template: Story<ButtonProps> = args => <Button {...args} />;
 
 export const Primary = Template.bind({});
